Change return types to int

+++ flightdataadapter.cpp
@@ -10,30 +10,30 @@ void FlightDataAdapter::SetFlightData(FlightData newData){
     emit flightDataChanged();
 }
 
-uint32_t FlightDataAdapter::TimeBootMs() const{
-    return data.TimeBootMs;
+int FlightDataAdapter::TimeBootMs() const{
+    return static_cast<int>(data.TimeBootMs);
 }
-int32_t FlightDataAdapter::Lat () const{
+int FlightDataAdapter::Lat () const{
 return data.Lat;
 }
-int32_t FlightDataAdapter::Lon () const{
+int FlightDataAdapter::Lon () const{
     return data.Lon;
 }
-int32_t FlightDataAdapter::Alt () const{
+int FlightDataAdapter::Alt () const{
     return data.Alt;
 }
-int32_t FlightDataAdapter::RelativeAlt () const{
+int FlightDataAdapter::RelativeAlt () const{
     return data.RelativeAlt;
 }
-int16_t FlightDataAdapter::Vx () const{
+int FlightDataAdapter::Vx () const{
     return data.Vx;
 }
-int16_t FlightDataAdapter::Vy () const{
+int FlightDataAdapter::Vy () const{
     return data.Vy;
 }
-int16_t FlightDataAdapter::Vz () const{
+int FlightDataAdapter::Vz () const{
     return data.Vz;
 }
-uint16_t FlightDataAdapter::Hdg () const{
-    return data.Hdg;
+int FlightDataAdapter::Hdg () const{
+    return static_cast<int>(data.Hdg);
 }
